+++ extensions/fluent-outdoors/CHANGELOG.md
@@ -1,9 +1,10 @@
 # Fluent Outdoors Changelog
 
-## [Initial Version] - 2023-01-30
+## [Initial Version] - 2023-01-31
 
 - Favourite tracks
 - Track listing per service
 - Track details with maintenance info
+- Sort tracks by name or maintenance info
 - Announcement listing per service
 - Announcement details with optional links